fix bugs

+++ src/api/playlists/handler.js
@@ -12,8 +12,8 @@ class PlaylistsHandler {
     async postPlaylistHandler (request, h) {
 
         this._validator.validatePlaylistsPayload(request.payload);
+        
         const { name } = request.payload;
-
         const { id: credentialId } = request.auth.credentials;
                         
         const playlistId = await this._service.addPlaylist({
@@ -34,29 +34,32 @@ class PlaylistsHandler {
     };
 
     async getPlaylistsHandler (request) {
+
         const { id: credentialId } = request.auth.credentials;
 
-        const playlists = await this._service.getPlaylists(credentialId);
+        const playlists = await this._service.getPlaylists( credentialId );
 
         return {
             status: 'success',
             data: {
                 playlists,
-            }
-        }
+            },
+        };
     };
 
-    async deletePlaylistHandler (request) {
+    async deletePlaylistByIdHandler (request) {
 
+        const { id: credentialId } = request.auth.credentials;       
         const { id } = request.params;
         
-        await this._service.deletePlaylistById(id);
+        await this._service.verifyPlaylistOwner( id, credentialId );
+        await this._service.deletePlaylistById( id );
 
         return {
             status: 'success',
             message: 'Playlist berhasil dihapus',
         };
     };
-}
+};
 
 module.exports = PlaylistsHandler;
